# documentation/API/openapi.yaml
openapi: 3.1.0
info:
  title: Gemini-CLI-Web API
  version: 0.1.0
  description: REST and WebSocket API for Gemini-CLI-Web
servers:
  - url: http://localhost:4008
    description: Express API (dev/prod)
  - url: http://localhost:4009
    description: Vite dev proxy
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GitStatus:
      type: object
      properties:
        branch: { type: string }
        modified: { type: array, items: { type: string } }
        added: { type: array, items: { type: string } }
        deleted: { type: array, items: { type: string } }
        untracked: { type: array, items: { type: string } }
    GitCommits:
      type: object
      properties:
        commits:
          type: array
          items:
            type: object
            properties:
              hash: { type: string }
              author: { type: string }
              email: { type: string }
              date: { type: string }
              message: { type: string }
              stats:
                type: object
                additionalProperties: true
    RemoteStatus:
      type: object
      properties:
        hasRemote: { type: boolean }
        branch: { type: string }
        remoteBranch: { type: string }
        remoteName: { type: string }
        ahead: { type: integer }
        behind: { type: integer }
        isUpToDate: { type: boolean }
security:
  - bearerAuth: []
paths:
...
  /api/git/status:
    get:
      summary: Git status
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GitStatus' }
  /api/git/branches:
    get:
      summary: Git branches
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/git/remote-status:
    get:
      summary: Remote status
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RemoteStatus' }
  /api/git/checkout:
    post:
      summary: Checkout branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
                branch: { type: string }
              required: [project, branch]
      responses:
        '200': { description: OK }
  /api/git/create-branch:
    post:
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
                branch: { type: string }
              required: [project, branch]
      responses:
        '200': { description: OK }
  /api/git/fetch:
    post:
      summary: Fetch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
              required: [project]
      responses:
        '200': { description: OK }
  /api/git/pull:
    post:
      summary: Pull
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
              required: [project]
      responses:
        '200': { description: OK }
        '500': { description: Error }
  /api/git/push:
    post:
      summary: Push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
              required: [project]
      responses:
        '200': { description: OK }
        '500': { description: Error }
  /api/git/discard:
    post:
      summary: Discard file changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
                file: { type: string }
              required: [project, file]
      responses:
        '200': { description: OK }
  /api/git/diff:
    get:
      summary: Diff for file
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
        - in: query
          name: file
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/git/commits:
    get:
      summary: Recent commits
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GitCommits' }
  /api/git/commit-diff:
    get:
      summary: Diff for commit
      parameters:
        - in: query
          name: project
          required: true
          schema: { type: string }
        - in: query
          name: commit
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/git/generate-commit-message:
    post:
      summary: Generate commit message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
                files:
                  type: array
                  items: { type: string }
              required: [project]
      responses:
        '200': { description: OK }
  /api/git/commit:
    post:
      summary: Commit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project: { type: string }
                message: { type: string }
                files:
                  type: array
                  items: { type: string }
              required: [project, message]
      responses:
        '200': { description: OK }